DeepSeek-Coder/tests/test_model_loader.py
import unittest
from deepseeker.models.model_loader import ModelLoader

class TestModelLoader(unittest.TestCase):

    def setUp(self):
        self.model_loader = ModelLoader()

    def test_load_model(self):
        # Assuming there is a method called load_model that loads a pre-trained model
        model = self.model_loader.load_model()
        self.assertIsNotNone(model, "The model should not be None after loading.")

    def test_model_is_singleton(self):
        # Assuming ModelLoader is implemented as a singleton to ensure one instance of the model
        model1 = self.model_loader.load_model()
        model2 = self.model_loader.load_model()
        self.assertIs(model1, model2, "ModelLoader should return the same instance of the model.")

    def test_model_predict(self):
        # Assuming the model has a predict method that takes input data
        model = self.model_loader.load_model()
        input_data = "Some input data for prediction"
        prediction = model.predict(input_data)
        self.assertIsNotNone(prediction, "The model should provide a prediction for the input data.")

if __name__ == '__main__':
    unittest.main()